Delete Admin controller, move endpoints to AdminUsers controller
